Move font awesome init
