+++ src/BloomBrowserUI/react_components/l10n.tsx
@@ -1,7 +1,6 @@
 import { CancelTokenStatic } from "axios";
 import * as React from "react";
 import theOneLocalizationManager from "../lib/localizationManager/localizationManager";
-import MarkdownIt = require("markdown-it");
 
 // This would be used by a control that doesn't have any text of its own,
 // but has children that need to be localized.
@@ -286,31 +285,6 @@ export class Label extends LocalizableElement<ILabelProps, ILocalizationState> {
             >
                 {this.getLocalizedContent()}
             </label>
-        );
-    }
-}
-
-// This component expects its content to be a single string (like all localizable elements) that
-// contains Markdown. It will convert that into HTML and show it.
-export class Markdown extends LocalizableElement<
-    ILabelProps,
-    ILocalizationState
-> {
-    public render() {
-        return (
-            <div
-                className={this.getClassName()}
-                onClick={() => {
-                    if (this.props.onClick) {
-                        this.props.onClick();
-                    }
-                }}
-                dangerouslySetInnerHTML={{
-                    __html: new MarkdownIt().render(
-                        this.state.translation || ""
-                    )
-                }}
-            />
         );
     }
 }
